Widen x86 userspace malloc tag to 64-bit, fix long-standing ax_check bug
- malloc.c/axiom.h: generation tag 32-bit -> 64-bit (unsigned long long -
  plain unsigned long is still 32-bit in -m32 code), xorshift64 PRNG with
  canonical shifts, 8-byte canary, ALIGN4 -> ALIGN8 - mirrors the same
  widening already done for src/kernel/heap.c and src/user/rv64/malloc.h.
- Halved SHADOW_COVER (32KB -> 16KB): shadow_tag[] is a static array baked
  into every one of the ~30 binaries statically linked against libaxiom;
  doubling the tag width without this blew the 1MB FAT12 image budget
  ("RM.BIN doesn't fit"). Net shadow footprint per binary is now slightly
  smaller than before the tag widening.
- memtest.c: updated to the new tag width.

Side effect: MEMTEST.BIN's long-standing "FAIL: ax_check live" (a fresh
allocation reporting as not-live) is now fixed - reproduced as ALL PASS
across 4 runs with different ASLR offsets. Root cause not fully isolated,
but forcing 8-byte alignment throughout (needed for the wider tag/canary)
most likely resolved a granule-alignment mismatch against the existing
8-byte SHADOW_GRANULE.

+++ src/libaxiom/include/axiom.h
@@ -91,13 +91,13 @@ void* ax_sbrk(int increment);           // сдвинуть heap break; (void*)-
 void* ax_malloc(unsigned int size);     // выделить size байт или NULL
 void  ax_free(void* ptr);               // освободить блок
 
-// Shadow + Memory Tagging API:
+// Shadow + Memory Tagging API (64-битный тег поколения):
 // ax_check      — 1 если [ptr,ptr+size) в состоянии OK (любой тег).
 // ax_alloc_tag  — тег текущего поколения блока (0 = freed/не выделен).
 // ax_check_tag  — 1 если OK И тег совпадает; ловит UAF после переиспользования.
-int          ax_check(void* ptr, unsigned int size);
-unsigned int ax_alloc_tag(void* ptr);
-int          ax_check_tag(void* ptr, unsigned int expected_tag, unsigned int size);
+int                ax_check(void* ptr, unsigned int size);
+unsigned long long ax_alloc_tag(void* ptr);
+int                ax_check_tag(void* ptr, unsigned long long expected_tag, unsigned int size);
 
 // Уровень stdio (реализован в stdio.c)
 void ax_putchar(char c);
